mongodb/requirements.txt
pymongo[srv]==4.6.2
python-dotenv==1.0.1